Rename the package to notifications.
This project was originally for sending ECR scan notifications but not
it's sending maintenance window events as well and we may use it for
other notifications so I've changed the package and jar names to
"notifications".
