+++ src/main/scala/uk/gov/nationalarchives/notifications/Lambda.scala
@@ -1,4 +1,4 @@
-package uk.gov.nationalarchives.scannotifications
+package uk.gov.nationalarchives.notifications
 
 import java.io.{InputStream, OutputStream}
 
@@ -6,10 +6,10 @@ import cats.FlatMap.ops.toAllFlatMapOps
 import cats.effect._
 import io.circe.parser.decode
 import messages.Messages._
-import uk.gov.nationalarchives.scannotifications.decoders.SSMMaintenanceDecoder.SSMMaintenanceEvent
-import uk.gov.nationalarchives.scannotifications.decoders.ScanDecoder.ScanEvent
-import uk.gov.nationalarchives.scannotifications.decoders._
-import uk.gov.nationalarchives.scannotifications.messages.EventMessages._
+import uk.gov.nationalarchives.notifications.decoders.SSMMaintenanceDecoder.SSMMaintenanceEvent
+import uk.gov.nationalarchives.notifications.decoders.ScanDecoder.ScanEvent
+import uk.gov.nationalarchives.notifications.decoders._
+import uk.gov.nationalarchives.notifications.messages.EventMessages._
 
 import scala.io.Source
